Implementa melhorias de acessibilidade e performance
