+++ src/pages/landingpage/index.html
@@ -3,11 +3,31 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="description" content="Curso de vestibular gratuito para estudantes de baixa renda que sonham com um ensino superior de qualidade.">
     <title>Einstein Floripa</title>
     <link rel="icon" href="/src/img/einstein-icon.svg" type="image/svg+xml"/>
-    <link href="/src/styles/output.css" rel="stylesheet" />
+
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap" rel="stylesheet">
+    
+    <link  href="/src/styles/output.css" rel="stylesheet"  media="print" onload="this.media='all'" />
+
 </head>
 <body id="body" class="bg-CREAM">
+    <header class="relative w-full px-auto">
+        <div class="flex flex-wrap items-center justify-between gap-5 px-8 py-5 w-full max-w-wrapper">
+            <a href="./" aria-label="Navegar para home" class="cursor-pointer w-24 md:w-28">
+                <img src="/src/img/einstein-logo.webp" alt="Logo escrito Einstein Floripa com desenho de rosto do Einstein, tudo em preto." height="45" width="126"/>
+            </a>
+            <button id="menu-button" aria-label="Menu" class="md:hidden cursor-pointer">
+                <i id="menu-icon" class="ph ph-list text-2xl"></i>
+            </button>
+            <nav id="nav" class="hidden absolute z-10 top-20 left-0 right-0 p-8 rounded-b-xl shadow-md border-t border-t-black border-opacity-10 bg-CREAM md:static md:z-auto md:top-auto md:left-auto md:right-auto md:p-0 md:rounded-none md:shadow-none md:border-0 md:bg-transparent md:flex md:w-full md:justify-center lg:w-fit">
+                <ul id="nav-bar-list" class="flex flex-col md:flex-row flex-wrap items-center gap-10"></ul>
+            </nav>
+        </div>
+    </header>
     <main id="main">
         <section class="flex flex-col items-center pt-8 md:pt-10 lg:pt-16">
             <div class="px-8">
@@ -24,9 +44,29 @@
                 </h2>
                 <div id="container-action-buttons-mobile-and-tablet" class="lg:hidden mt-6 flex flex-col md:flex-row gap-4 w-full"></div>
             </div>
-            <img src="/src/img/aprovados.png" alt="Grupo de estudantes calouros de 2024, posando para foto em clima de celebração. Todos têm os nomes dos cursos escritos nos braços, como 'Sistemas', 'Psico', 'Jornal', entre outros, e alguns seguram placas com as palavras 'Calouro 2024' e 'UFSC'. A expressão deles é de alegria e animação." width="1100" height="700"/>
+            <img src="/src/img/aprovados.webp" alt="Grupo de estudantes calouros de 2024, posando para foto em clima de celebração. Todos têm os nomes dos cursos escritos nos braços, como 'Sistemas', 'Psico', 'Jornal', entre outros, e alguns seguram placas com as palavras 'Calouro 2024' e 'UFSC'. A expressão deles é de alegria e animação." width="1100" height="567" layout="responsive"/>
         </section>
         <section id="about" class="bg-BLUE-50">
+            <div class="w-full bg-BLUE-300">
+                <ul class="flex flex-row flex-wrap gap-10 justify-between items-center px-8 py-3.5 max-w-wrapper">
+                    <li class="flex flex-row items-center gap-3 text-CREAM">
+                        <span class="font-bebas text-6xl">+100</span>
+                        <span class="font-roboto text-xl font-medium capitalize">aprovados</span>
+                    </li>
+                    <li class="flex flex-row items-center gap-3 text-CREAM">
+                        <span class="font-bebas text-6xl">+200</span>
+                        <span class="font-roboto text-xl font-medium capitalize">vidas transformadas</span>
+                    </li>
+                    <li class="flex flex-row items-center gap-3 text-CREAM">
+                        <span class="font-bebas text-6xl">+1260</span>
+                        <span class="font-roboto text-xl font-medium capitalize">horas de aula</span>
+                    </li>
+                    <li class="flex flex-row items-center gap-3 text-CREAM">
+                        <span class="font-bebas text-6xl">+4500</span>
+                        <span class="font-roboto text-xl font-medium capitalize">horas trabalhadas</span>
+                    </li>
+                </ul>
+            </div>
             <div class="px-10 py-12 lg:py-28 flex flex-col lg:flex-row items-center justify-center gap-10 lg:gap-24">
                 <div id="introduction">
                     <h2 class="title mb-4">
@@ -47,11 +87,11 @@
                     
                 </div>
                 <div class="relative">
-                    <img src="/src/img/polaroids.png" alt="Colagem de três fotos polaroides: a primeira mostra um grupo de pessoas sorrindo e celebrando, com a legenda 'Equipe - 2023'; a segunda, uma pessoa de costas segurando um cartaz com o texto 'De Aluno para Aluno - Einstein Floripa', em um evento ao ar livre; e a terceira, duas pessoas colando post-its em um cartaz que diz 'Einstein em uma palavra', durante o 'Piquenique dos Aprovados 2024'." width="593" height="540"/>
+                    <img src="/src/img/polaroids.webp" alt="Colagem de três fotos polaroides: a primeira mostra um grupo de pessoas sorrindo e celebrando, com a legenda 'Equipe - 2023'; a segunda, uma pessoa de costas segurando um cartaz com o texto 'De Aluno para Aluno - Einstein Floripa', em um evento ao ar livre; e a terceira, duas pessoas colando post-its em um cartaz que diz 'Einstein em uma palavra', durante o 'Piquenique dos Aprovados 2024'." width="593" height="540"/>
                     <div id="socials-cta" class="hidden md:block absolute bottom-0"></div>
                 </div>
             </div>
-            <img src="/src/img/bubbles-divider.png" alt="Um padrão horizontal de semicírculos azuis claros agindo como divisor.">
+            <img src="/src/img/bubbles-divider.webp" alt="Um padrão horizontal de semicírculos azuis claros agindo como divisor." width="1440" height="40">
         </section> 
         <section id="testimonials" class="md:px-10 py-12 lg:py-28">
             <h2 class="title px-8 md:px-0 mt-10 mb-4 text-center">
@@ -60,7 +100,6 @@
             </h2>
         </section>    
     </main>
-    
     <script type="module" src="/src/pages/landingpage/page.js"></script>
 </body>
 </html>
