+++ src/styles/output.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap" rel="stylesheet');
-
 @import url('https://unpkg.com/@phosphor-icons/web@2.1.1/src/regular/style.css');
 
 @import url('https://unpkg.com/@phosphor-icons/web@2.1.1/src/fill/style.css');
@@ -837,8 +835,8 @@ video {
   border-bottom-left-radius: 0.75rem;
 }
 
-.border{
-  border-width: 1px;
+.border-2{
+  border-width: 2px;
 }
 
 .border-b{
@@ -885,6 +883,11 @@ video {
 .bg-BLUE-100{
   --tw-bg-opacity: 1;
   background-color: rgb(182 216 254 / var(--tw-bg-opacity));
+}
+
+.bg-BLUE-300{
+  --tw-bg-opacity: 1;
+  background-color: rgb(120 143 232 / var(--tw-bg-opacity));
 }
 
 .bg-BLUE-50{
@@ -1174,12 +1177,12 @@ video {
   --tw-bg-opacity: 0.1;
 }
 
-.hover\:opacity-90:hover{
-  opacity: 0.9;
-}
-
 .hover\:opacity-80:hover{
   opacity: 0.8;
+}
+
+.hover\:opacity-90:hover{
+  opacity: 0.9;
 }
 
 .hover\:after\:w-full:hover::after{
